Implement changes in PR comments
Use Distance and Angle enums to specify units instead of a string. This
is better type safety and removes the need to check arguments in convert.
Remove conversion tables and just put units per reference unit as an
enum field.
Multiple by to / from in conversion instead of from / to.