Link Android SDK docs in javadoc
Download and link Android SDK docs in javadoc for the Canvas class,
which is used a single time in SampleProcessor as part of the signature
of a method we don't even implement.
Finally figure out how to add the Android SDK to the javadoc classpath
because otherwise that single class usage breaks the build.